+++ lib/sysdef.f90
@@ -7,7 +7,7 @@
 !-----------------------------------------------------------------------
 ! Parameters common to all trajectories
 !-----------------------------------------------------------------------
-    integer*8                                   :: natm,nsta,nintraj,&
+    integer                                     :: natm,nsta,nintraj,&
                                                    nstep
     real*8, dimension(:), allocatable           :: alpha,atnum,atmass
     real*8                                      :: dt,tf
